+++ dataweaver/apps/web/src/components/foundations/toaster/config.ts
@@ -1,0 +1,14 @@
+/** How long a toast stays on screen before auto-dismissing. */
+export const TOAST_AUTO_DISMISS_DELAY = 3000;
+
+/** Vertical gap: the peek when stacked and the spacing when expanded. */
+export const ROW_GAP = 16;
+
+/** How much each toast behind the front one shrinks. */
+export const SCALE_STEP = 0.025;
+
+/** How many toasts stay visible in the collapsed stack. */
+export const MAX_VISIBLE_TOASTS = 3;
+
+/** How far below its resting spot a new toast starts before sliding in. */
+export const ENTER_OFFSET = 24;
